Extract InstructionFlags etc. into flags.py
